Replace ContextLoader HOC with ContextProvider using React.createContext
- Create ContextProvider component with createContext + provider pattern
- Extract useAppContext hook (using React 19's `use`) into context.ts
- Consume context via useAppContext in App and Header, eliminating prop
  drilling of `context` through App → Header
- Use a ref guard to prevent duplicate /api/context fetches in StrictMode,
  fixing the TODO about 4 calls on startup
- Remove the legacy ContextLoader HOC and the contextState workaround
