+++ app/frontend/src/main.tsx
@@ -3,15 +3,15 @@ import { createRoot } from "react-dom/client";
 import "bootstrap/dist/css/bootstrap.min.css";
 import "./styles/globals.css";
 import App from "./components/App";
-import ContextLoader from "./components/ContextLoader";
-
-const AppWithContext = ContextLoader(App);
+import ContextProvider from "./components/ContextProvider";
 
 const rootElement = document.getElementById("root");
 if (!rootElement) throw new Error("Root element not found");
 
 createRoot(rootElement).render(
   <StrictMode>
-    <AppWithContext />
+    <ContextProvider>
+      <App />
+    </ContextProvider>
   </StrictMode>,
 );
